+++ collector-amazon-ivs/src/main/java/com/bitmovin/analytics/amazon/ivs/AmazonIvsUtil.kt
@@ -1,0 +1,10 @@
+package com.bitmovin.analytics.amazon.ivs
+
+class AmazonIvsUtil {
+    companion object {
+        fun getUserAgent(): String {
+//            TODO not implemented
+            return ""
+        }
+    }
+}